pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-extract-kit"
version = "0.1.0"
authors = [
    {name = "Bin Wang", email = "ictwangbin@gmail.com"}
]
description = "A Comprehensive Toolkit for High-Quality PDF Content Extraction."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "pypdf2",
    "matplotlib",
    "pyyaml>=6.0.0",
    "frontend",
    "pymupdf",
    "opencv-python>=4.6.0",
    "ultralytics>=8.2.85",
    "rootutils>=1.0.7",
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "aiofiles>=23.1.0",
    "omegaconf>=2.3.0",
    "doclayout-yolo==0.0.3",
    "unimernet==0.2.1",
    "paddlepaddle>=3.0.0",
    "paddleocr==2.7.3",
    "modelscope>=1.25.0",
    "pdf2image>=1.16.3",
    "pillow>=9.5.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.2",
    "nougat-ocr>=0.1.18",
    "table-transformer>=0.1.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "pytest",
]

[project.optional-dependencies]
layout_detection = [
    "transformers"
]
formula_detection = [
    "ultralytics"
]
